Portrait Cropping/requirements.txt
pillow==9.4.0
unitypy==1.9.26
